# src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mosify Reward System API
  version: 1.0.0
servers:
  - url: '/'
paths:
...
  /boards/{id}/users/{userId}:
    post:
      summary: Add an existing user to a board
      operationId: addUserToBoard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: alias
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User added to board successfully
        '404':
          description: Board or User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove a user from a board
      operationId: removeUserFromBoard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User removed from board successfully
        '404':
          description: Board-User membership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users:
    get:
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete user by ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/boards:
    get:
      summary: Get boards belonging to a user
      operationId: getUserBoards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of boards the user belongs to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /tasks:
    get:
      summary: Get all tasks
      operationId: getAllTasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
    post:
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /tasks/{id}:
    get:
      summary: Get task by ID
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete task by ID
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{id}/execute:
    post:
      summary: Execute task for a user
      operationId: executeTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task executed successfully and transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Business validation error (insufficient balance or already completed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User or Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UserRequest:
      type: object
      required:
        - name
        - username
        - password
      properties:
        name:
          type: string
          example: "Oscar"
        username:
          type: string
          example: "oscar"
        password:
          type: string
          example: "mysecurepassword"
    UserResponse:
      type: object
      required:
        - id
        - name
        - username
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        username:
          type: string
...
    TaskRequest:
      type: object
      required:
        - title
        - categoryId
        - type
        - pointsValue
      properties:
        title:
          type: string
          example: "Study math for 1 hour"
        categoryId:
          type: string
          format: uuid
        type:
          type: string
          enum: [SINGLE_USE, RECURRENT]
          example: "RECURRENT"
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY]
          nullable: true
          example: "DAILY"
        pointsValue:
          type: integer
          example: 50
        active:
          type: boolean
          example: true
    TaskResponse:
      type: object
      required:
        - id
        - title
        - categoryId
        - type
        - pointsValue
        - active
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        categoryId:
          type: string
          format: uuid
        type:
          type: string
          enum: [SINGLE_USE, RECURRENT]
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY]
          nullable: true
        pointsValue:
          type: integer
        active:
          type: boolean
    TransactionResponse:
      type: object
      required:
        - id
        - userId
        - pointsAffected
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
          nullable: true
        pointsAffected:
          type: integer
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "INSUFFICIENT_BALANCE"
        message:
          type: string
          example: "The user has insufficient points balance."
...
    BoardResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...